improves : structure of project 	1.renames core.models -> core.time_models 	2.puts model Goal in core.graph_models 	3.renames core.classes -> core.time_utils 	4.puts to_microseconds() function in core.time_utils 	5.removes escapeerrands.timeutils 	6.registers Goal model to admin
