+++ core/time_utils.py
@@ -1,9 +1,13 @@
 from datetime import datetime as dt, timedelta as td
-from django.utils import timezone as tz
+
+
+def to_microseconds(time_delta):
+    if isinstance(time_delta, time_delta):
+        return time_delta.days * 86400000000 + time_delta.seconds * 1000000 + time_delta.microseconds
 
 
 class Stub:
-    def __init__(self, epoch=tz.now(), end=tz.now(), duration=td()):
+    def __init__(self, epoch=dt.now(), end=dt.now(), duration=td()):
         self.epoch = epoch
         self.end = end
         self.duration = duration
